+++ RTMPHaishinKit/Tests/RTMP/RTMPTimestampTests.swift
@@ -46,13 +46,18 @@ import Testing
             .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.000)),
             .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.020)),
             .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.057)),
-            .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.077))
+            .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.077)),
+            .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.114)),
+            .init(hostTime: AVAudioTime.hostTime(forSeconds: 100.134))
         ]
         let aacPacketDuration = 1024.0 / 44100.0
         var timestamp = RTMPTimestamp<AVAudioTime>()
         #expect(try timestamp.update(times[0], preferredDelta: aacPacketDuration) == 0)
         #expect(try timestamp.update(times[1], preferredDelta: aacPacketDuration) == 23)
         #expect(try timestamp.update(times[2], preferredDelta: aacPacketDuration) == 23)
-        #expect(try timestamp.update(times[3], preferredDelta: aacPacketDuration) == 24)
+        #expect(try timestamp.update(times[3], preferredDelta: aacPacketDuration) == 23)
+        #expect(try timestamp.update(times[4], preferredDelta: aacPacketDuration) == 23)
+        #expect(try timestamp.update(times[5], preferredDelta: aacPacketDuration) == 24)
+        #expect(abs(timestamp.updatedAt - (100.0 + aacPacketDuration * 5)) < 0.001)
     }
 }